+++ packages/backend/test/modules/artworks/artworks.repository.spec.ts
@@ -1,12 +1,12 @@
 import { DataSource, In, Repository } from 'typeorm';
 
+import { Language } from '@/src/common/enums/language.enum';
 import { ArtworksRepository } from '@/src/modules/artworks/artworks.repository';
 import { ArtworkTranslation } from '@/src/modules/artworks/entities/artwork-translations.entity';
 import { Artwork } from '@/src/modules/artworks/entities/artworks.entity';
 import { Platform } from '@/src/modules/artworks/enums/platform.enum';
 import { Sort } from '@/src/modules/artworks/enums/sort-type.enum';
 import { Genre } from '@/src/modules/genres/entities/genres.entity';
-import { Language } from '@/src/modules/genres/enums/language.enum';
 import { GenresRepository } from '@/src/modules/genres/genres.repository';
 import { ArtworkTranslationsFactory } from '@/test/factories/artwork-translations.factory';
 import { ArtworksFactory } from '@/test/factories/artworks.factory';
