+++ packages/backend/test/modules/artworks/validators/status.validator.spec.ts
@@ -1,7 +1,7 @@
+import { Language } from '@/src/common/enums/language.enum';
 import { Artwork } from '@/src/modules/artworks/entities/artworks.entity';
 import { StatusValidator } from '@/src/modules/artworks/validators/status.validator';
 import { Genre } from '@/src/modules/genres/entities/genres.entity';
-import { Language } from '@/src/modules/genres/enums/language.enum';
 import { ArtworkTranslationsFactory } from '@/test/factories/artwork-translations.factory';
 import { ArtworksFactory } from '@/test/factories/artworks.factory';
 import { GenreTranslationsFactory } from '@/test/factories/genre-translations.factory';
