+++ packages/backend/test/modules/genres/genres.controller.spec.ts
@@ -4,6 +4,7 @@ import request from 'supertest';
 import { DataSource, In, Repository } from 'typeorm';
 
 import { PAGE_SIZE } from '@/src/common/constants/page-size.constant';
+import { Language } from '@/src/common/enums/language.enum';
 import { Artwork } from '@/src/modules/artworks/entities/artworks.entity';
 import { AuthService } from '@/src/modules/auth/auth.service';
 import { Administrator } from '@/src/modules/auth/entities/administrator.entity';
@@ -14,7 +15,6 @@ import { DeleteGenresDto } from '@/src/modules/genres/dtos/delete-genres.dto';
 import { UpdateGenreDto } from '@/src/modules/genres/dtos/update-genre.dto';
 import { GenreTranslation } from '@/src/modules/genres/entities/genre-translations.entity';
 import { Genre } from '@/src/modules/genres/entities/genres.entity';
-import { Language } from '@/src/modules/genres/enums/language.enum';
 import { GenreErrorCode } from '@/src/modules/genres/exceptions/genres.exception';
 import { GenresController } from '@/src/modules/genres/genres.controller';
 import { GenresMapper } from '@/src/modules/genres/genres.mapper';
